refactor(test): Move Castings.java to testclasses (#2155)
